moving things into app component
